refactor(asciidoc): move tests into internal file
